+++ scripts/macros/2024/spells/dustDevil.js
@@ -1,4 +1,0 @@
-import {dustDevil as dustDevilLegacy} from '../../../legacyMacros.js';
-export let dustDevil = {
-    ...dustDevilLegacy,
-};